remove unnecessary cast

+++ java/src/jmri/util/swing/multipane/TwoPaneTBWindow.java
@@ -139,13 +139,13 @@ abstract public class TwoPaneTBWindow extends jmri.util.JmriJFrame {
         if(title!=null){
             if(fontSize<=4)
                 fontSize = title.getFont().getSize()-1;
-            title.setFont(title.getFont().deriveFont(Font.PLAIN,(float)fontSize));
+            title.setFont(title.getFont().deriveFont(Font.PLAIN,fontSize));
             statusItemPanel.add(title);
         }
         if(value!=null){
             if(fontSize<=4)
                 fontSize = value.getFont().getSize()-1;
-            value.setFont(value.getFont().deriveFont(Font.PLAIN,(float)fontSize));
+            value.setFont(value.getFont().deriveFont(Font.PLAIN,fontSize));
             statusItemPanel.add(value);
         }
         addToStatusBox(statusItemPanel);
